+++ src-tauri/Cargo.toml
@@ -3,7 +3,7 @@ name = "neko_todo"
 version = "0.1.5"
 description = "A Tauri App"
 authors = ["you"]
-edition = "2021"
+edition = "2024"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
